auth-service/requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14
django-cors-headers>=4.3
PyJWT>=2.8
psycopg2-binary>=2.9
